refactor: winner를 구하는 함수에 대한 클래스 생성 및 수정
